requirements.txt
pandas
matplotlib
cartopy
tqdm
numpy
seaborn
pyyaml